java/user-group-security/api-server/src/test/java/io/stackdocker/ugs/apiserver/mapper/UserMapperTest.java
package io.stackdocker.ugs.apiserver.mapper;

import io.stackdocker.ugs.apiserver.mapper.UserMapper;
import io.fairymagic.core.ugs.domain.User;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.test.context.junit4.SpringRunner;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

@RunWith(SpringRunner.class)
@MybatisTest
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
public class UserMapperTest {

    @Autowired
    private UserMapper userMapper;

    @Test
    public void addOne() {
        User user = new User();
        long id = userMapper.addOne(user);
        assertThat(id > 0);
    }
}
